src/requirements.txt
aiogram==2.20
python_dotenv
outline-vpn-api
